+++ CHANGELOG.md
@@ -27,7 +27,10 @@ and this project adheres to [PEP 440 Versioning](https://peps.python.org/pep-044
 - **Pipeline python backend**:
   - Provide POC version. <!-- TG-2051, TG-2074 -->
   - Add a `Swagger` link to the published component to open the service's 'Swagger UI'. <!-- TG-2098 -->
-
+- **Documentation**:
+  - Cross-links w/ youwol's symbols in docstrings more tolerant. <!-- TG 2128 -->
+  - Add `admin/system/documentation-check` end point. <!-- TG 2129 -->
+  
 ### Changed
 
 - **Pipeline Typescript**: